Add browser specific main
