+++ src/main-browser.js
@@ -1,0 +1,5 @@
+
+module.exports = require('./main-common');
+
+exports.homepage = '{{package-homepage}}';
+exports.version = '{{package-version}}';
