+++ src/main-common.js
@@ -1,6 +1,4 @@
 
-var environment = require('./environment');
-
 var DiffPatcher = require('./diffpatcher').DiffPatcher;
 exports.DiffPatcher = DiffPatcher;
 
@@ -46,20 +44,3 @@ exports.clone = function() {
   }
   return defaultInstance.clone.apply(defaultInstance, arguments);
 };
-
-
-if (environment.isBrowser) {
-  exports.homepage = '{{package-homepage}}';
-  exports.version = '{{package-version}}';
-} else {
-  var packageInfoModuleName = '../package.json';
-  var packageInfo = require(packageInfoModuleName);
-  exports.homepage = packageInfo.homepage;
-  exports.version = packageInfo.version;
-
-  var formatterModuleName = './formatters';
-  var formatters = require(formatterModuleName);
-  exports.formatters = formatters;
-  // shortcut for console
-  exports.console = formatters.console;
-}
