+++ src/main.js
@@ -1,0 +1,20 @@
+
+module.exports = require('./main-common');
+
+var environment = require('./environment');
+
+if (environment.isBrowser) {
+  exports.homepage = '{{package-homepage}}';
+  exports.version = '{{package-version}}';
+} else {
+  var packageInfoModuleName = '../package.json';
+  var packageInfo = require(packageInfoModuleName);
+  exports.homepage = packageInfo.homepage;
+  exports.version = packageInfo.version;
+
+  var formatterModuleName = './formatters';
+  var formatters = require(formatterModuleName);
+  exports.formatters = formatters;
+  // shortcut for console
+  exports.console = formatters.console;
+}
